+++ CHANGELOG.md
@@ -50,6 +50,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   to get pushed to gerrit.
   [#8568](https://github.com/jj-vcs/jj/issues/8568)
 
+* `jj file track --include-ignored` now works when `fsmonitor.backend="watchman"`.
+  [#8427](https://github.com/jj-vcs/jj/issues/8427)
+
 ## [0.37.0] - 2026-01-07
 
 ### Release highlights